--- __mocks__/react-adaptive-hooks/index.js
@@ -1,13 +1,0 @@
-const useNetworkStatus = () => {
-  return { effectiveConnectionType: '4g' };
-};
-
-const useMemoryStatus = () => {
-  return { deviceMemory: 3 };
-};
-
-const useHardwareConcurrency = () => {
-  return { numberOfLogicalProcessors: 4 };
-};
-
-export { useNetworkStatus, useHardwareConcurrency, useMemoryStatus };
