+++ src/__mocks__/use-hardware-concurrency.js
@@ -1,0 +1,5 @@
+const useHardwareConcurrency = () => {
+    return { numberOfLogicalProcessors: 4 };
+  };
+
+  export {useHardwareConcurrency}
