+++ src/__mocks__/use-memory-status.js
@@ -1,0 +1,5 @@
+const useMemoryStatus = () => {
+    return { deviceMemory: 3 };
+  };
+
+  export {useMemoryStatus}
