+++ src/__mocks__/use-network-status.js
@@ -1,0 +1,5 @@
+const useNetworkStatus = () => {
+    return { effectiveConnectionType: '4g' };
+  };
+  
+  export {useNetworkStatus}
